delete and restore for expenses , delete image from upload folder

+++ back_end/controllers/recycleController.js
@@ -4,41 +4,44 @@ const Worker = require('../models/Worker')
 const Memo = require('../models/Memo')
 const Expense = require('../models/Expense')
 
-exports.getDeletedSites = async (req, res, next) => {
-  try{
-    let query = {isDeleted : true};
+const fs = require("fs");
+const path = require("path");
 
-    if(req.user.role === 'admin'){
+exports.getDeletedSites = async (req, res, next) => {
+  try {
+    let query = { isDeleted: true };
+
+    if (req.user.role === 'admin') {
       query.createdBy = req.user.userId;
     }
 
     const deletedSites = await Site.find(query);
     res.json(deletedSites)
-  } catch(err){
+  } catch (err) {
     res.status(500).json({ message: "Error fetching deleted sites", error: err.message });
   }
 }
 
 exports.getDeletedManagers = async (req, res, next) => {
-  try{
-    let query = {isDeleted : true};
+  try {
+    let query = { isDeleted: true };
 
-    if(req.user.role === 'admin'){
+    if (req.user.role === 'admin') {
       query.createdBy = req.user.userId;
     }
 
     const deletedManagers = await Manager.find(query);
     res.json(deletedManagers)
-  } catch(err){
+  } catch (err) {
     res.status(500).json({ message: "Error fetching deleted managers", error: err.message });
   }
 }
 
 exports.getDeletedWorkers = async (req, res, next) => {
   try {
-    const {siteId} = req.params;
+    const { siteId } = req.params;
 
-    let query = {isDeleted : true};
+    let query = { isDeleted: true };
     query.site = siteId;
 
     const deletedWorkers = await Worker.find(query)
@@ -50,9 +53,9 @@ exports.getDeletedWorkers = async (req, res, next) => {
 
 exports.getDeletedMemos = async (req, res, next) => {
   try {
-    const {siteId} = req.params;
+    const { siteId } = req.params;
 
-    let query = {isDeleted : true};
+    let query = { isDeleted: true };
     query.siteId = siteId;
 
     const deletedMemos = await Memo.find(query)
@@ -64,9 +67,9 @@ exports.getDeletedMemos = async (req, res, next) => {
 
 exports.getDeletedExpenses = async (req, res, next) => {
   try {
-    const {siteId} = req.params;
+    const { siteId } = req.params;
 
-    let query = {isDeleted : true};
+    let query = { isDeleted: true };
     query.siteId = siteId;
 
     const deletedExpenses = await Expense.find(query)
@@ -78,11 +81,21 @@ exports.getDeletedExpenses = async (req, res, next) => {
 
 exports.deleteSite = async (req, res, next) => {
   try {
-    const {siteId} = req.params;
+    const { siteId } = req.params;
+
+    const site = await Site.findById(siteId);
+    if (!site) return res.status(404).json({ message: "Site not found" });
+
+    if (site && site.siteImage) {
+      const Path = path.join(__dirname, "../uploads/sites", site.siteImage)
+      if (fs.existsSync(Path)) {
+        fs.unlinkSync(Path)
+      }
+    }
 
     const deletedSite = await Site.findByIdAndDelete(siteId)
     res.json(deletedSite)
-    
+
   } catch (err) {
     res.status(500).json({ message: "Error deleting site permanently", error: err.message });
   }
@@ -90,19 +103,19 @@ exports.deleteSite = async (req, res, next) => {
 
 exports.restoreSite = async (req, res, next) => {
   try {
-    const {siteId} = req.params;
+    const { siteId } = req.params;
 
     const restoredSite = await Site.findByIdAndUpdate(
       siteId,
       {
-        isDeleted : false,
-        deletedAt : null
+        isDeleted: false,
+        deletedAt: null
       },
-      {new : true}
+      { new: true }
 
     )
     res.json(restoredSite)
-    
+
   } catch (err) {
     res.status(500).json({ message: "Error restoring deleted site", error: err.message });
   }
@@ -110,11 +123,21 @@ exports.restoreSite = async (req, res, next) => {
 
 exports.deleteManager = async (req, res, next) => {
   try {
-    const {managerId} = req.params;
+    const { managerId } = req.params;
+
+    const manager = await Manager.findById(managerId);
+    if (!manager) return res.status(404).json({ message: "manager not found" });
+
+    if (manager && manager.managerImage) {
+      const Path = path.join(__dirname, "../uploads/managers", manager.managerImage)
+      if (fs.existsSync(Path)) {
+        fs.unlinkSync(Path)
+      }
+    }
 
     const deletedManager = await Manager.findByIdAndDelete(managerId)
     res.json(deletedManager)
-    
+
   } catch (err) {
     res.status(500).json({ message: "Error deleting manager permanently", error: err.message });
   }
@@ -122,20 +145,62 @@ exports.deleteManager = async (req, res, next) => {
 
 exports.restoreManager = async (req, res, next) => {
   try {
-    const {managerId} = req.params;
+    const { managerId } = req.params;
 
     const restoredManager = await Manager.findByIdAndUpdate(
       managerId,
       {
-        isDeleted : false,
-        deletedAt : null
+        isDeleted: false,
+        deletedAt: null
       },
-      {new : true}
+      { new: true }
 
     )
     res.json(restoredManager)
-    
+
   } catch (err) {
     res.status(500).json({ message: "Error restoring deleted manager", error: err.message });
+  }
+}
+
+exports.deleteExpense = async (req, res, next) => {
+  try {
+    const { expenseId } = req.params;
+
+    const expense = await Expense.findById(expenseId);
+    if (!expense) return res.status(404).json({ message: "Expense not found" });
+
+    if (expense && expense.billImage) {
+      const Path = path.join(__dirname, "../uploads/bills", expense.billImage)
+      if (fs.existsSync(Path)) {
+        fs.unlinkSync(Path)
+      }
+    }
+
+    const deletedExpense = await Expense.findByIdAndDelete(expenseId)
+    res.json(deletedExpense)
+
+  } catch (err) {
+    res.status(500).json({ message: "Error deleting Expense permanently", error: err.message });
+  }
+}
+
+exports.restoreExpense = async (req, res, next) => {
+  try {
+    const { expenseId } = req.params;
+
+    const restoredExpense = await Expense.findByIdAndUpdate(
+      expenseId,
+      {
+        isDeleted: false,
+        deletedAt: null
+      },
+      { new: true }
+
+    )
+    res.json(restoredExpense)
+
+  } catch (err) {
+    res.status(500).json({ message: "Error restoring deleted expense", error: err.message });
   }
 }
